requirements.txt
adal==1.0.2
applicationinsights==0.11.5
argcomplete==1.9.4
asn1crypto==0.24.0
astroid==1.6.5
azure-batch==4.1.3
azure-cli==2.0.41
azure-cli-acr==2.0.28
azure-cli-acs==2.1.3
azure-cli-advisor==0.5.1
azure-cli-ams==0.1.1
azure-cli-appservice==0.1.36
azure-cli-backup==1.1.2
azure-cli-batch==3.2.4
azure-cli-batchai==0.3.1
azure-cli-billing==0.1.8
azure-cli-cdn==0.0.14
azure-cli-cloud==2.0.15
azure-cli-cognitiveservices==0.1.13
azure-cli-command-modules-nspkg==2.0.1
azure-cli-configure==2.0.17
azure-cli-consumption==0.3.1
azure-cli-container==0.2.1
azure-cli-core==2.0.41
azure-cli-cosmosdb==0.1.21
azure-cli-dla==0.1.0
azure-cli-dls==0.0.22
azure-cli-dms==0.0.1
azure-cli-eventgrid==0.1.12
azure-cli-eventhubs==0.1.3
azure-cli-extension==0.1.0
azure-cli-feedback==2.1.3
azure-cli-find==0.2.11
azure-cli-interactive==0.3.25
azure-cli-iot==0.1.21
azure-cli-keyvault==2.0.23
azure-cli-lab==0.0.22
azure-cli-monitor==0.1.8
azure-cli-network==2.1.5
azure-cli-nspkg==3.0.2
azure-cli-profile==2.0.27
azure-cli-rdbms==0.2.5
azure-cli-redis==0.2.14
azure-cli-reservations==0.2.1
azure-cli-resource==2.0.32
azure-cli-role==2.0.26
azure-cli-servicebus==0.1.2
azure-cli-servicefabric==0.0.12
azure-cli-sql==2.0.27
azure-cli-storage==2.0.35
azure-cli-vm==2.0.35
azure-common==1.1.13
azure-datalake-store==0.0.22
azure-graphrbac==0.40.0
azure-keyvault==1.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.0.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-compute==4.0.0rc2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.0.0
azure-mgmt-containerregistry==2.0.0
azure-mgmt-containerservice==4.1.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==2.0.0
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.0.0rc1
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.0.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.0.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-managementgroups==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0rc1
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.0.0rc3
azure-mgmt-nspkg==2.0.0
azure-mgmt-rdbms==1.2.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.0.0rc2
azure-mgmt-servicebus==0.5.1
azure-mgmt-servicefabric==0.1.0
azure-mgmt-sql==0.9.1
azure-mgmt-storage==1.5.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-multiapi-storage==0.2.1
azure-nspkg==2.0.0
azure-storage-blob==1.3.0
azure-storage-common==1.3.0
azure-storage-nspkg==3.0.0
backports.functools-lru-cache==1.5
bcrypt==3.1.4
beautifulsoup4==4.6.3
blinker==1.4
boto3==1.7.54
botocore==1.10.54
bs4==0.0.1
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
configparser==3.5.0
cryptography==2.2.2
docutils==0.14
dparse==0.4.1
elastic-apm==2.2.1
entrypoints==0.2.3
enum34==1.1.6
Flask==1.0.2
funcsigs==1.0.2
futures==3.1.1
gunicorn==19.9.0
httplib2==0.11.3
humanfriendly==4.13
idna==2.6
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
json-logging-py==0.2
keyring==13.2.1
knack==0.4.1
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
msrest==0.5.2
msrestazure==0.4.34
oauth2client==4.1.2
oauthlib==2.1.0
packaging==17.1
paramiko==2.4.1
pbr==4.1.0
prompt-toolkit==2.0.3
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
pydocumentdb==2.3.2
Pygments==2.2.0
PyJWT==1.6.4
pymongo==3.7.0
PyNaCl==1.2.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
python-dotenv==0.8.2
python-json-logger==0.1.9
python3-logstash==0.4.72
pytz==2018.5
PyYAML==4.2b4
raven==6.9.0
requests==2.19.1
requests-futures==0.9.7
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.13
safety==1.8.1
schedule==0.5.0
scp==0.11.0
singledispatch==3.4.0.3
six==1.11.0
sshtunnel==0.1.4
tabulate==0.8.2
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.23
vsts-cd-manager==1.0.2
wcwidth==0.1.7
websocket-client==0.48.0
Werkzeug==0.14.1
Whoosh==2.7.4
wrapt==1.10.11
xmltodict==0.11.0
